Create 54-day.py
Steps Followed

In the first line of code we have imported counter from collections.
As the problem statement suggests we need to find the count of words. So for that, we have created a list of words.
In the next step we convert that list to a unique list using the counter module in python. 
Next in the step we printed the count of unique words in the list using len method.
Lastly we .values() method to find and print all the values of words inside res variable